refactor: remove source deduplicated logic
